Move get_commit_diff to utils
Since the method is static and may also be useful elsewhere, utils.repo
is a better place than LocalProject. Currently it is only used in packit
and hence is safe to move.
