+++ packit/api.py
@@ -54,7 +54,12 @@ from packit.upstream import Upstream
 from packit.utils import commands
 from packit.utils.changelog_helper import ChangelogHelper
 from packit.utils.extensions import assert_existence
-from packit.utils.repo import shorten_commit_hash, get_next_commit, commit_exists
+from packit.utils.repo import (
+    shorten_commit_hash,
+    get_next_commit,
+    commit_exists,
+    get_commit_diff,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -312,7 +317,7 @@ class PackitAPI:
         distro_path = self.up.local_project.working_dir / DISTRO_DIR
         for commit in self.dg.local_project.get_commits(revision_range, reverse=True):
             commits.append(commit)
-            diffs.append(self.dg.local_project.get_commit_diff(commit))
+            diffs.append(get_commit_diff(commit))
             for diff in diffs[-1]:
                 if diff.a_path == "sources" or diff.b_path == "sources":
                     raise PackitException(
